QRT macro is removed an replaced where possible with the sqr function (macro expansion would cause multiple calculation of provided complex expressions?), also olx_abs function is added to replace fabs. smatd now is a mat3i and vec3d association. Three reflection mergerers introduced : ShelxMerger, StandardMerger, and UnitMerger. A bug fixed - default initialisation of TReflection::Tag was missing.

+++ sdl/tptrlist.h
@@ -232,7 +232,7 @@ public:
 #endif
     T *D = Items[from];
     if( from > to )  {
-      for( int i=from-1; i >= to; i-- )
+      for( size_t i =from-1; i >= to; i-- )
       Items[i+1] = Items[i];
     }
     else  {
